+++ src/main/java/org/jenkinsci/plugins/pipeline/utility/steps/fs/FileWrapper.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.jenkinsci.plugins.workflow.utility.steps.fs;
+package org.jenkinsci.plugins.pipeline.utility.steps.fs;
 
 import hudson.FilePath;
 import org.jenkinsci.plugins.scriptsecurity.sandbox.whitelists.Whitelisted;
