+++ src/main/resources/index.jelly
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?jelly escape-by-default='true'?>
 <div>
-    Utility steps for workflows.
+    Utility steps for pipeline jobs.
 </div>
