+++ src/main/resources/org/jenkinsci/plugins/pipeline/utility/steps/zip/ZipStep/help-archive.html
